requirements.txt
autopep8==1.4.4
cached-property==1.5.1
Django==2.2.6
django-filter==2.2.0
django-url-filter==0.3.13
djangorestframework==3.10.3
enum-compat==0.0.2
numpy==1.17.2
pandas==0.25.1
psycopg2==2.8.3
pycodestyle==2.5.0
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
